Split of RBOSS from BOSS, moved the BOSSIndividual class into its own file.
